pyproject.toml
[project]
name = "kittycrew"
version = "0.1.0"
description = "Cute A2A-powered crew scheduler for Claude Code, Codex, and GitHub Copilot."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn>=0.30",
    "jinja2>=3.1",
    "httpx>=0.27",
    "pydantic>=2.7",
    "a2a-sdk[http-server]>=0.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]

[project.scripts]
kittycrew = "kittycrew.main:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
kittycrew = ["static/*.js", "static/*.css", "static/avatars/*.svg", "templates/*.html"]

[tool.pytest.ini_options]
testpaths = ["tests"]
